Logo on Figma
We had originally tried to make the logo on Canva, but it was not conducive to the logo-making since it does not have a measurement tool. Anthony and Dalond then recommended using Figma to make this logo instead - it can allow for lines to be the same number of pixels/at the same angle. Dalond and I then worked together to implement this vision in Figma.